Ghosts are good.
Loading and stuff.

Rewrote `_cursor` to use the index into the leaf page generated by the
`Descent` object, which required the `Descent` object to be a little
more descerning about the page types, whether or not they were leaf
pages or branch pages. The `key` function generator now generates a
function that will use the `page.ghosts` low index in its call to `find`
when the page is a leaf page. The `descend` function does not logically
and its `index` property if the page is a leaf page. The `leftMost`
function returns `ghosts` if the page is a leaf page.

When recording a delete, we don't need a ghosts flag. We'll know. But,
we won't know who's a ghost in the positions array, so we've added
ghostlness to that journal entry.

Loading a leaf page will increment a page's `ghosts` property if the
first record is deleted.

Calling insert objects journal entires now in the documentation.

Setting balancer recorded length accounting for ghosts.

Oddly, calling key in `_cursor` was passing in an array, not the first
element of the array. How did that ever work?

Removed the `first` property because we'll know the left most leaf page
by its address; `-1`.

Went through all of the JSON fixtures and change the right sibling page
to `0` where it was `-1`.